Add prompt issue tool
