+++ libs/labelbox/src/labelbox/schema/ontology_building/tool_type.py
@@ -4,6 +4,7 @@ from enum import Enum
 class ToolType(Enum):
     STEP_REASONING = "step-reasoning"
     FACT_CHECKING = "fact-checking"
+    PROMPT_ISSUE = "prompt-issue"
 
     @classmethod
     def valid(cls, tool_type: str) -> bool:
