--- libs/labelbox/src/labelbox/schema/tool_building/__init__.py
@@ -1,5 +1,0 @@
-import labelbox.schema.tool_building.tool_type
-import labelbox.schema.tool_building.variant
-import labelbox.schema.tool_building.step_reasoning_tool
-import labelbox.schema.tool_building.fact_checking_tool
-import labelbox.schema.tool_building.tool_type_mapping
