+++ libs/labelbox/tests/integration/conftest.py
@@ -22,6 +22,13 @@ from labelbox import (
     Tool,
 )
 from labelbox.schema.data_row import DataRowMetadataField
+from labelbox.schema.ontology_building.fact_checking_tool import (
+    FactCheckingTool,
+)
+from labelbox.schema.ontology_building.prompt_issue_tool import PromptIssueTool
+from labelbox.schema.ontology_building.step_reasoning_tool import (
+    StepReasoningTool,
+)
 from labelbox.schema.ontology_kind import OntologyKind
 from labelbox.schema.user import User
 
@@ -581,6 +588,7 @@ def chat_evaluation_ontology(client, rand_gen):
             ),
             StepReasoningTool(name="step reasoning"),
             FactCheckingTool(name="fact checking"),
+            PromptIssueTool(name="prompt issue"),
         ],
         classifications=[
             Classification(
