+++ libs/labelbox/tests/integration/test_ontology.py
@@ -5,8 +5,12 @@ import pytest
 
 from labelbox import MediaType, OntologyBuilder, Tool
 from labelbox.orm.model import Entity
-from labelbox.schema.tool_building.fact_checking_tool import FactCheckingTool
-from labelbox.schema.tool_building.step_reasoning_tool import StepReasoningTool
+from labelbox.schema.ontology_building.fact_checking_tool import (
+    FactCheckingTool,
+)
+from labelbox.schema.ontology_building.step_reasoning_tool import (
+    StepReasoningTool,
+)
 
 
 def test_feature_schema_is_not_archived(client, ontology):
